dtrace: validate argument pointer to d_path()
When an invalid pointer was being passed to d_path(), the system could
crash with a OOPS.  This was the result of the kernel implementation
(reasonably) expecting the pointer to be referencing a valid path
struct.  We now validate the argument passed to d_path() against the
paths for files known to the current task.

Orabug: 21304207

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Acked-by: Nick Alcock <nick.alcock@oracle.com>